requirements.txt
Django==3.0
mysqlclient==2.0.3
djangorestframework==3.12.4
Pillow==8.4.0
qiniu==7.5.0